requirements.txt
openai~=0.27.2
python-socketio==5.8.0
fastapi~=0.95.0
uvicorn[standard]~=0.21.1
mysql-connector-python~=8.0.32